requirements.txt
langgraph
langchain
langchain-community
langchain-google-genai
langsmith
python-dotenv
ipykernel
jupyterlab
rich
pandas
pypdf
langchain-text-splitters
langchain-chroma 
langchain-openai